Middle-layer model refactor.

+++ test/v1/fixtures/taxableIncome/detail/selfEmployments/summary/LossClaimSummaryFixtures.scala
@@ -16,33 +16,33 @@
 
 package v1.fixtures.taxableIncome.detail.selfEmployments.summary
 
-import play.api.libs.json.{JsObject, JsValue, Json}
+import play.api.libs.json.{ JsObject, JsValue, Json }
 import v1.models.response.getCalculation.taxableIncome.detail.selfEmployment.summary.LossClaimsSummary
 
 object LossClaimSummaryFixtures {
 
-  val selfEmploymentId: String                                 = "AAIS12345678904"
-  val incomeSourceType: String                               = "01"
-  val incomeSourceName: String                               = "abcdefghijklm"
-  val totalIncome: BigDecimal                                    = 79291394
-  val totalExpenses: BigDecimal                                  = 89005890
-  val netProfit: BigInt                                      = 93480427
-  val netLoss: BigInt                                        = 10017816
-  val totalAdditions: BigInt                                 = 39901282
-  val totalDeductions: BigInt                                = 80648172
-  val accountingAdjustments: BigDecimal                      = -8769926.99
-  val taxableProfit: BigInt                                  = 92149284
-  val adjustedIncomeTaxLoss: BigDecimal                      = 2
-  val totalBroughtForwardIncomeTaxLosses: Option[BigDecimal] = Some(1)
-  val lossForCSFHL: BigDecimal                               = 2
-  val broughtForwardIncomeTaxLossesUsed: Option[BigDecimal]  = Some(2)
-  val taxableProfitAfterLossesDeduction: BigDecimal = 2
-  val totalIncomeTaxLossesCarriedForward: Option[BigDecimal] = Some(3)
-  val class4Loss: BigDecimal                                 = 2
-  val totalBroughtForwardClass4Losses: Option[BigDecimal]    = Some(1)
-  val broughtForwardClass4LossesUsed: Option[BigDecimal]     = Some(2)
-  val carrySidewaysClass4LossesUsed: Option[BigDecimal]      = Some(2)
-  val totalClass4LossesCarriedForward: Option[BigDecimal]    = Some(3)
+  val selfEmploymentId: String                           = "AAIS12345678904"
+  val incomeSourceType: String                           = "01"
+  val incomeSourceName: String                           = "abcdefghijklm"
+  val totalIncome: BigDecimal                            = 79291394
+  val totalExpenses: BigDecimal                          = 89005890
+  val netProfit: BigInt                                  = 93480427
+  val netLoss: BigInt                                    = 10017816
+  val totalAdditions: BigInt                             = 39901282
+  val totalDeductions: BigInt                            = 80648172
+  val accountingAdjustments: BigDecimal                  = -8769926.99
+  val taxableProfit: BigInt                              = 92149284
+  val adjustedIncomeTaxLoss: BigInt                      = 2
+  val totalBroughtForwardIncomeTaxLosses: Option[BigInt] = Some(1)
+  val lossForCSFHL: BigDecimal                           = 2
+  val broughtForwardIncomeTaxLossesUsed: Option[BigInt]  = Some(2)
+  val taxableProfitAfterLossesDeduction: BigInt          = 2
+  val totalIncomeTaxLossesCarriedForward: Option[BigInt] = Some(3)
+  val class4Loss: BigDecimal                             = 2
+  val totalBroughtForwardClass4Losses: Option[BigInt]    = Some(1)
+  val broughtForwardClass4LossesUsed: Option[BigInt]     = Some(2)
+  val carrySidewaysClass4LossesUsed: Option[BigInt]      = Some(2)
+  val totalClass4LossesCarriedForward: Option[BigInt]    = Some(3)
 
   val lossClaimsSummaryResponse = LossClaimsSummary(
     totalBroughtForwardIncomeTaxLosses,
